requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.8
SQLAlchemy==2.0.12
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
requests==2.30.0
bcrypt==4.0.1
aiofiles==23.1.0
loguru==0.7.0
# AssemblyAI SDK - SDK officiel pour l'API AssemblyAI
assemblyai==0.37.0
